create-pet-view-cliente

+++ database/migrations/2022_09_06_233627_create_pesos_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('pet_id');
             $table->foreign('pet_id')->references('id')->on('pets');
-            $table->integer('peso');
+            $table->double('peso');
             $table->softDeletes();
             $table->timestamps();
         });
